# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ReuseHub API
  version: 0.3.0
  description: |
    REST API for ReuseHub inventory management, community discovery,
    authentication, profile privacy, and reuse-interest analysis.

    Browser clients authenticate with the HTTP-only `reusehub_token` cookie.
    Native/mobile clients and API tools may instead send the same JWT through
    `Authorization: Bearer <token>`. Protected operations accept either method.
servers:
  - url: https://YOUR_DOMAIN/api
    description: Production API; replace YOUR_DOMAIN after DNS and HTTPS are configured
  - url: http://localhost:5000/api
    description: Local development API
...
paths:
...
  /auth/verify-email:
    post:
      tags: [Authentication]
      operationId: verifyEmail
      summary: Verify a local account
      description: |
        Verifies the one-time email token, creates the browser session cookie,
        and returns an access token for native/mobile clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: Email verified and authenticated session created
          headers:
            Set-Cookie:
              description: HTTP-only browser session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          description: Invalid, expired, or malformed verification token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /auth/google:
    post:
      tags: [Authentication]
      operationId: googleLogin
      summary: Log in with a Google ID token
      description: |
        Verifies a Google Identity Services credential server-side, creates or
        updates the local ReuseHub account, sets the browser cookie, and returns
        an access token for native/mobile clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleLoginRequest"
      responses:
        "200":
          description: Google login successful
          headers:
            Set-Cookie:
              description: HTTP-only browser session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Google credential could not be verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Google authentication is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/forgot-password:
    post:
      tags: [Authentication]
      operationId: forgotPassword
      summary: Request a password-reset email
      description: Returns the same generic response whether or not the account exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          description: Generic password-reset response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/reset-password:
    post:
      tags: [Authentication]
      operationId: resetPassword
      summary: Reset a password with a one-time token
      description: |
        Changes the password, marks the email verified, sets the browser session
        cookie, and returns an access token for native/mobile clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset and authenticated session created
          headers:
            Set-Cookie:
              description: HTTP-only browser session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          description: Invalid, expired, or malformed reset token/password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: reusehub_token
      description: HTTP-only JWT cookie used by the React web client
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned as accessToken for Flutter, native, Postman, and Swagger clients
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: Too many authentication requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          nullable: true
          description: Validation details may be included outside production
...
    GoogleLoginRequest:
      type: object
      required: [credential]
      properties:
        credential:
          type: string
          minLength: 20
          description: Google Identity Services ID-token credential
    EmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    TokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 20
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          minLength: 20
        password:
          $ref: "#/components/schemas/StrongPassword"
    StrongPassword:
      type: string
      minLength: 10
      maxLength: 128
      pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9]).+$'
      description: Must contain lowercase, uppercase, and numeric characters
    AuthSessionResponse:
      type: object
      required: [accessToken, user]
      properties:
        message:
          type: string
        accessToken:
          type: string
          description: JWT for Bearer authentication by native/mobile and API clients
        user:
          $ref: "#/components/schemas/User"
    ForgotPasswordResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: If the account exists, a reset link has been sent.
        developmentResetUrl:
          type: string
          format: uri
          description: Present only when a development email preview is available
        resetToken:
          type: string
          description: Present only in the automated test environment
...
    User:
      type: object
      required:
        [id, firstName, lastName, email, schoolDomain, communityName, contact, emailVerified]
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        schoolDomain:
          type: string
        communityName:
          type: string
        contact:
          $ref: "#/components/schemas/ContactSettings"
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
...
    ContactSettings:
      type: object
      required: [emailVisible, phone, phoneVisible, preferredContact]
      properties:
        emailVisible:
          type: boolean
        phone:
          type: string
          maxLength: 30
        phoneVisible:
          type: boolean
        preferredContact:
          type: string
          enum: [email, phone]
...